+++ R/summary_cross2.R
@@ -59,6 +59,12 @@ function(cross2)
 ind_ids <-
 function(cross2)
 {
+    # going to sneak in use for calc_genoprob (or fst_genoprob)
+    # this is probably really bad form
+    if(inherits(cross2, "calc_genoprob")) {
+        return(dimnames(cross2)[[1]])
+    }
+
     if(!is.cross2(cross2))
         stop('Input cross must have class "cross2"')
     unique( c(ind_ids_geno(cross2), ind_ids_pheno(cross2), ind_ids_covar(cross2)) )